+++ CHANGELOG.md
@@ -40,6 +40,7 @@ New features:
 * core env: working lxd prototype
 
 Fix:
+* file.upload: honor sudo
 * db.database: mysql/mariadb default to latin1
 * status: more expressive tests
 * file.stat: handle empty stdout